adding a data_structures folder in preparation
