+++ source/databricks/calculation_engine/package/calculation/energy/aggregators/grid_loss_aggregators.py
@@ -16,7 +16,9 @@ from pyspark.sql import DataFrame
 
 import package.calculation.energy.aggregators.transformations as t
 from package.calculation.energy.energy_results import EnergyResults
-from package.calculation.preparation.grid_loss_responsible import GridLossResponsible
+from package.calculation.preparation.data_structures.grid_loss_responsible import (
+    GridLossResponsible,
+)
 from package.codelists import (
     MeteringPointType,
     QuantityQuality,
