+++ source/databricks/calculation_engine/package/calculation/energy/aggregators/metering_point_time_series_aggregators.py
@@ -18,7 +18,7 @@ from package.calculation.energy.energy_results import EnergyResults
 from package.calculation.energy.aggregators.transformations.aggregate_sum_and_quality import (
     aggregate_quantity_and_quality,
 )
-from package.calculation.preparation.quarterly_metering_point_time_series import (
+from package.calculation.preparation.data_structures.quarterly_metering_point_time_series import (
     QuarterlyMeteringPointTimeSeries,
 )
 from package.codelists import (
