+++ source/databricks/calculation_engine/package/calculation/energy/hour_to_quarter.py
@@ -14,10 +14,10 @@
 
 import pyspark.sql.functions as f
 
-from package.calculation.preparation.prepared_metering_point_time_series import (
+from package.calculation.preparation.data_structures.prepared_metering_point_time_series import (
     PreparedMeteringPointTimeSeries,
 )
-from package.calculation.preparation.quarterly_metering_point_time_series import (
+from package.calculation.preparation.data_structures.quarterly_metering_point_time_series import (
     QuarterlyMeteringPointTimeSeries,
 )
 from package.codelists import MeteringPointResolution
