+++ source/databricks/calculation_engine/package/calculation/output/basis_data_factory.py
@@ -16,7 +16,7 @@ from pyspark.sql import DataFrame
 from pyspark.sql.functions import col
 
 from package.calculation.CalculationResults import BasisDataContainer
-from package.calculation.preparation.prepared_metering_point_time_series import (
+from package.calculation.preparation.data_structures.prepared_metering_point_time_series import (
     PreparedMeteringPointTimeSeries,
 )
 from package.calculation.preparation.transformations import basis_data
@@ -71,7 +71,6 @@ def _get_ga_basis_data(
     timeseries_quarter_df: DataFrame,
     timeseries_hour_df: DataFrame,
 ) -> tuple[DataFrame, DataFrame, DataFrame]:
-
     timeseries_quarter_df = timeseries_quarter_df.drop(
         BasisDataColname.energy_supplier_id
     ).withColumnRenamed(BasisDataColname.grid_area, PartitionKeyName.GRID_AREA)
@@ -97,7 +96,6 @@ def _get_es_basis_data(
     timeseries_quarter_df: DataFrame,
     timeseries_hour_df: DataFrame,
 ) -> tuple[DataFrame, DataFrame, DataFrame]:
-
     timeseries_quarter_df = timeseries_quarter_df.withColumnRenamed(
         BasisDataColname.energy_supplier_id, PartitionKeyName.ENERGY_SUPPLIER_GLN
     ).withColumnRenamed(BasisDataColname.grid_area, PartitionKeyName.GRID_AREA)
