+++ source/databricks/calculation_engine/package/calculation/preparation/data_structures/charge_link_metering_point_periods.py
