+++ source/databricks/calculation_engine/package/calculation/preparation/data_structures/grid_loss_responsible.py
