+++ source/databricks/calculation_engine/package/calculation/preparation/data_structures/input_charges.py
@@ -13,12 +13,16 @@
 # limitations under the License.
 from dataclasses import dataclass
 
-from package.calculation.preparation.prepared_subscriptions import PreparedSubscriptions
-from package.calculation.preparation.prepared_tariffs import PreparedTariffs
+from pyspark.sql import DataFrame
+
+from package.calculation.preparation.data_structures.charge_master_data import (
+    ChargeMasterData,
+)
+from package.calculation.preparation.data_structures.charge_prices import ChargePrices
 
 
 @dataclass
-class PreparedChargesContainer:
-    hourly_tariffs: PreparedTariffs | None = None
-    daily_tariffs: PreparedTariffs | None = None
-    subscriptions: PreparedSubscriptions | None = None
+class InputChargesContainer:
+    charge_master_data: ChargeMasterData
+    charge_prices: ChargePrices
+    charge_links: DataFrame
