+++ source/databricks/calculation_engine/package/calculation/preparation/data_structures/prepared_metering_point_time_series.py
