+++ source/databricks/calculation_engine/package/calculation/preparation/data_structures/quarterly_metering_point_time_series.py
