+++ source/databricks/calculation_engine/package/calculation/preparation/transformations/metering_point_time_series.py
@@ -16,7 +16,7 @@ from pyspark.sql import DataFrame
 import pyspark.sql.functions as f
 from pyspark.sql.types import DecimalType
 
-from package.calculation.preparation.prepared_metering_point_time_series import (
+from package.calculation.preparation.data_structures.prepared_metering_point_time_series import (
     PreparedMeteringPointTimeSeries,
 )
 from package.constants import Colname
