+++ source/databricks/calculation_engine/package/calculation/wholesale/get_wholesale_metering_point_time_series.py
@@ -14,7 +14,7 @@
 import pyspark.sql.functions as f
 
 from package.calculation.energy.energy_results import EnergyResults
-from package.calculation.preparation.prepared_metering_point_time_series import (
+from package.calculation.preparation.data_structures.prepared_metering_point_time_series import (
     PreparedMeteringPointTimeSeries,
 )
 from package.codelists import (
