+++ source/databricks/calculation_engine/package/calculation/wholesale/wholesale_calculation.py
@@ -13,8 +13,6 @@
 # limitations under the License.
 
 
-from pyspark.sql import DataFrame
-
 import package.calculation.output.wholesale_storage_model_factory as factory
 import package.calculation.wholesale.tariff_calculators as tariffs
 import package.calculation.wholesale.tariff_calculators as tariff_calculator
@@ -22,9 +20,15 @@ import package.calculation.wholesale.subscription_calculators as subscription_ca
 
 from ..CalculationResults import WholesaleResultsContainer
 from ..calculator_args import CalculatorArgs
-from ..preparation.prepared_charges import PreparedChargesContainer
-from ..preparation.prepared_subscriptions import PreparedSubscriptions
-from ..preparation.prepared_tariffs import PreparedTariffs
+from package.calculation.preparation.data_structures.prepared_charges import (
+    PreparedChargesContainer,
+)
+from package.calculation.preparation.data_structures.prepared_subscriptions import (
+    PreparedSubscriptions,
+)
+from package.calculation.preparation.data_structures.prepared_tariffs import (
+    PreparedTariffs,
+)
 from ...codelists import AmountType
 from ...infrastructure import logging_configuration
