+++ source/databricks/calculation_engine/tests/calculation/preparation/transformations/prepared_metering_point_time_series_factory.py
@@ -16,7 +16,7 @@ from decimal import Decimal
 
 from pyspark.sql import Row, SparkSession
 
-from package.calculation.preparation.prepared_metering_point_time_series import (
+from package.calculation.preparation.data_structures.prepared_metering_point_time_series import (
     PreparedMeteringPointTimeSeries,
     prepared_metering_point_time_series_schema,
 )
